Use int32 shard tensor conversion for torch compatibility
Some pod torch builds reject numpy uint16 in torch.from_numpy. Convert shard arrays to int32 to keep token values intact while improving cross-environment compatibility.

Made-with: Cursor

+++ train_gpt.py
@@ -531,7 +531,8 @@ def load_data_shard(file: Path) -> Tensor:
     tokens_np = np.fromfile(file, dtype="<u2", count=num_tokens, offset=header_bytes)
     if tokens_np.size != num_tokens:
         raise ValueError(f"Short read for {file}")
-    return torch.from_numpy(tokens_np.astype(np.uint16, copy=False))
+    # Some torch builds reject uint16 numpy conversion; int32 is widely supported.
+    return torch.from_numpy(tokens_np.astype(np.int32, copy=False))
 
 
 class TokenStream:
